id domain auto

+++ api/models/domain.py
@@ -2,6 +2,6 @@ from django.db import models
 
 
 class Domain(models.Model):
-    domain_id = models.BigIntegerField(primary_key=True)
+    domain_id = models.AutoField(primary_key=True)
     name = models.CharField(max_length=50)
     description = models.TextField()
